Make lab instructions better suited for printing
If a learner wants to print out lab instructions, or save them to
a PDF, then it really doesn't make sense to restrict their height.
Also, it doesn't really help if we print out the contents of the
Guacamole terminal.

Thus, restrict the height limitation to @media screen, and suppress
the terminal window's display with @media print.